+++ squash_wip.go
@@ -1,8 +1,8 @@
 package main
 
 import (
-	config "github.com/remotemobprogramming/mob/v4/configuration"
-	"github.com/remotemobprogramming/mob/v4/say"
+	config "github.com/remotemobprogramming/mob/v5/configuration"
+	"github.com/remotemobprogramming/mob/v5/say"
 	"io"
 	"os"
 	"path/filepath"